requirements-minimal.txt
# Extended requirements for full app functionality
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2
python-dotenv==1.0.0
requests==2.31.0
urllib3==2.1.0
anthropic==0.34.0
groq==0.8.0
python-multipart==0.0.6
python-dateutil==2.8.2
httpx==0.25.2
requests==2.31.0
